Handle bare and trailing where separately
`g:haskell_indent_after_bare_where` controls the following (using a
value of 2 for this example):

```
f x = y
  -- a bare where follows
  where
  >>y = 0
```

This change makes this rule stricter by only treating the `where`
keyword as bare if it appears alone on the line, with only some
allowance for whitespace. (Not even comments are allowed.)

On top of that `where` is now treated as trailing if it appears last on
a line. `g:haskell_indent_after_trailing_where` now controls the
subsequent indentation. For instance with a value of 3:

```
-- a trailing where follows
instance Show MyData where
>>>show (MyData x) = "MyData"
```